chronicle: env name change
