+++ chronicle/Cargo.toml
@@ -42,7 +42,6 @@ tracing-core = "0.1.28"
 opentelemetry = { version = "0.28.0", default-features = false, features = ["trace"] }
 opentelemetry_sdk = { version = "0.28.0", default-features = false, features = ["trace"] }
 opentelemetry-otlp = { version = "0.28.0", features = ["metrics", "grpc-tonic"] }
-dhat = "0.3.3"
 
 [dev-dependencies]
 ed25519-dalek = "2.1.1"
